object_detection_app/requirements.txt
Django==3.2.7
Pillow==8.3.1
torch==1.9.0
yolov5==5.0.8
